+++ lang/src_ref/line_col.rs
@@ -5,9 +5,9 @@
 #[derive(Clone, Debug, Default)]
 pub struct LineCol {
     /// Line number (1..)
-    pub line: u32,
+    pub line: usize,
     /// Column number (1..)
-    pub col: u32,
+    pub col: usize,
 }
 
 impl std::fmt::Display for LineCol {
